requirements.txt
-e .
aws-cdk.core
aws-cdk.aws_iam
aws-cdk.aws_ec2
